zenity loop, partial sync

+++ music.bash_aliases
@@ -4,7 +4,10 @@
 
 unalias youtube-dl 2>/dev/null
 alias youtube-dl='youtube-dl --proxy socks5://127.0.0.1:1080/ --force-ipv4'
+
 TEMPLATE='%(id)s.%(ext)s'
+
+SEP='^'
 
 function sub {
   if [ "$#" -eq 1 ]; then
@@ -24,9 +27,19 @@ function sub {
 
 function sync {
   echo
-  adb shell rm -v -- /sdcard/Music/\*
+  adb devices
   echo
-  adb push -- *.lrc *.m4a *.opus /sdcard/Music/
+  if [ "$#" -ne 0 ]; then
+    for i in "$@"; do
+      adb push -- "$i" /sdcard/Music/
+    done
+  else
+    echo -n "Remove all audios from Android, "
+    confirm
+    adb shell rm -v -- /sdcard/Music/\*
+    echo
+    adb push -- *.lrc *.m4a *.opus /sdcard/Music/
+  fi
   echo
   echo "Please Scan again in Musicolet"
   echo
@@ -47,13 +60,8 @@ function conv {
   realpath -- "$2"
 }
 
-function aud {
-  if [ "$#" -lt 1 ] || [ "$#" -gt 3 ]; then
-    echo
-    echo "  ${FUNCNAME[0]} <ID> [TITLE [ARTIST]]"
-    echo
-    return
-  fi
+function download_convert {
+  test "$#" -eq 3 || return
   youtube-dl --output "$TEMPLATE" -f bestaudio -- "$1"
   WEBM="$1.webm"
   OPUS="$1.opus"
@@ -67,4 +75,40 @@ function aud {
     echo "unsupported audio container"
     return
   fi
+}
+
+function confirm {
+  echo -n "OK? "
+  read -r
+}
+
+function aud_loopbody {
+  test "$#" -eq 0 || return 1
+  OUTPUT="$(zenity \
+    --forms \
+    --title='Audio from YouTube' \
+    --window-icon=question \
+    --add-entry="ID" \
+    --add-entry="TITLE" \
+    --add-entry="ARTIST" \
+    --text="character '$SEP' not allowed in any field" \
+    --separator="$SEP" \
+  )"
+  R="$?"
+  test "$R" -eq 0 || return 1
+  ID="$(awk -F "$SEP" '{print $1}' <<<"$OUTPUT")"
+  TITLE="$(awk -F "$SEP" '{print $2}' <<<"$OUTPUT")"
+  ARTIST="$(awk -F "$SEP" '{print $3}' <<<"$OUTPUT")"
+  echo "ID=$ID"
+  echo "TITLE=$TITLE"
+  echo "ARTIST=$ARTIST"
+  confirm
+  download_convert "$ID" "$TITLE" "$ARTIST"
+}
+
+function aud {
+  while true; do
+    aud_loopbody
+    test "$?" -eq 0 || return
+  done
 }
